Unifying event

+++ hftbacktest/src/backtest/assettype.rs
@@ -1,10 +1,10 @@
 /// Calculates the value amount and the equity according to the asset type.
 pub trait AssetType {
     /// Calculates the value amount.
-    fn amount(&self, price: f32, qty: f32) -> f64;
+    fn amount(&self, price: f64, qty: f64) -> f64;
 
     /// Calculates the equity.
-    fn equity(&self, price: f32, balance: f64, position: f64, fee: f64) -> f64;
+    fn equity(&self, price: f64, balance: f64, position: f64, fee: f64) -> f64;
 }
 
 /// The common type of asset where the contract's notional value is linear to the quote currency.
@@ -21,12 +21,12 @@ impl LinearAsset {
 }
 
 impl AssetType for LinearAsset {
-    fn amount(&self, exec_price: f32, qty: f32) -> f64 {
-        self.contract_size * exec_price as f64 * qty as f64
+    fn amount(&self, exec_price: f64, qty: f64) -> f64 {
+        self.contract_size * exec_price * qty
     }
 
-    fn equity(&self, price: f32, balance: f64, position: f64, fee: f64) -> f64 {
-        balance + self.contract_size * position * price as f64 - fee
+    fn equity(&self, price: f64, balance: f64, position: f64, fee: f64) -> f64 {
+        balance + self.contract_size * position * price - fee
     }
 }
 
@@ -44,11 +44,11 @@ impl InverseAsset {
 }
 
 impl AssetType for InverseAsset {
-    fn amount(&self, exec_price: f32, qty: f32) -> f64 {
-        self.contract_size * qty as f64 / exec_price as f64
+    fn amount(&self, exec_price: f64, qty: f64) -> f64 {
+        self.contract_size * qty / exec_price
     }
 
-    fn equity(&self, price: f32, balance: f64, position: f64, fee: f64) -> f64 {
-        -balance - self.contract_size * position / price as f64 - fee
+    fn equity(&self, price: f64, balance: f64, position: f64, fee: f64) -> f64 {
+        -balance - self.contract_size * position / price - fee
     }
 }
